[RFR] Fix facets creation after publication
