+++ src/api/controller/api/publish.spec.js
@@ -9,9 +9,9 @@ import {
     addTransformResultToDoc,
     versionTransformResult,
     publishCharacteristics,
-    publishFacets,
 } from './publish';
 import getDocumentTransformer from '../../services/getDocumentTransformer';
+import publishFacets from './publishFacets';
 
 describe('publish', () => {
     describe('doPublish', () => {
@@ -142,37 +142,6 @@ describe('publish', () => {
             expect(ctx.publishedCharacteristic.addNewVersion).toHaveBeenCalledWith({
                 transformed: 'document',
             });
-        });
-    });
-
-    describe('publishFacets', () => {
-        const ctx = {
-            publishedDataset: {
-                findDistinctValuesForField: createSpy().andReturn(Promise.resolve(['value1', 'value2'])),
-                countByFacet: createSpy().andReturn(Promise.resolve(100)),
-            },
-            publishedFacet: {
-                insertFacet: createSpy().andReturn(Promise.resolve()),
-            },
-        };
-        const facetFields = [{
-            name: 'facet1',
-        }, {
-            name: 'facet2',
-        }];
-
-        before(async () => {
-            await publishFacets(ctx, facetFields);
-        });
-
-        it('should call publishedDataset.findDistinctValuesForField for each facet field', () => {
-            expect(ctx.publishedDataset.findDistinctValuesForField).toHaveBeenCalledWith('facet1');
-            expect(ctx.publishedDataset.findDistinctValuesForField).toHaveBeenCalledWith('facet2');
-        });
-
-        it('should call ctx.publishedFacet.insertFacet for each facet field with their distinct values', () => {
-            expect(ctx.publishedFacet.insertFacet).toHaveBeenCalledWith('facet1', [{ value: 'value1', count: 100 }, { value: 'value2', count: 100 }]);
-            expect(ctx.publishedFacet.insertFacet).toHaveBeenCalledWith('facet2', [{ value: 'value1', count: 100 }, { value: 'value2', count: 100 }]);
         });
     });
